Rename Abstract{Newton,FinDiff}L10nData
